# util/swagger/doc.yaml
swagger: '2.0'
info:
  description: LunchRun
  version: 1.0.0
  title: Lunch Run
  
host: 54.180.89.60
basePath: /
  
paths:
  /auth/signup:
    post:
      summary: 회원가입
      tags: 
      - "Auth"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/user"
      responses:
        200:
          description: "{userId : 회원가입된 유저의 아이디}"
        409:
          description: '중복된 회원 정보로 가입 시도시 :   {"message" : "email || phone"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "email || phone || alias || phone",
                                    "location": "body"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
          
  /auth/login:
    post:
      summary: 로그인 토큰 받아오기
      tags: 
      - "Auth"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: object
          properties:
            email:
              type: string

            password:
              type: string

      responses:
        200:
          description: "{token : 생성된 토큰}"
        403:
          description: '회원 정보에 맞지 않은 값으로 로그인 시도시 : {"message" : "login failed || token failed"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "email || password",
                                    "location": "body"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
          
  /auth/check:
    post:
      summary: 로그인 토큰 체크
      tags: 
      - "Auth"
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string

      responses:
        200:
          description: '{"userInfo": {회원 정보} }'
        403:
          description: '토큰이 없거나 토큰이 맞지 않을때 : {"message" : "Not Token || invalid token"}'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'

  /restaurantCategory:
    get:
      summary: 식당 카테고리 정보 리스트
      tags:
      - restaurant_category
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      responses:
        200:
          description: "카테고리 리스트"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/restaurant_category"
        403:
          description: '{"message" : "Not Token || invalid token"}'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []

  /restaurant_point:
    get:
      summary: 식당 지도 포인트 정보
      tags:
      - restaurant
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      - in: "body"
        name: "body"
        required: false
        type: object
        properties:
          category_id:
            type: integer
      responses:
        200:
          description: '[{"id": **,"lat": **,"lng": ***,"category_id": *}]'
        403:
          description: '{"message" : "Not Token || invalid token"}'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []
          
  /restaurant:
    get:
      summary: 식당 리스트
      tags:
      - restaurant
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      - in: "body"
        name: "body"
        required: false
        type: object
        properties:
          category_id:
            type: integer
          page:
            type: integer
      responses:
        200:
          description: '[{"id": **, "name": "***", "floor": **, "introduction": "http://place.map.kakao.com/******", "lat": ***, "lng": ***, "address": "*****", "address_road": "*****", "category_id": *}]'
        403:
          description: '{"message" : "Not Token || invalid token"}'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []

    post:
      summary: 식당 정보 입력
      tags:
      - restaurant
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/restaurant"
      responses:
        200:
          description: "{restaurantId : 입력된 식당의 아이디}"
        403:
          description: '{"message" : "Not Token || invalid token"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "name || floor || lat || lng || category_id",
                                    "location": "body"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []
        
    put:
      summary: 식당 정보 수정
      tags:
      - restaurant
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/restaurant"
      responses:
        200:
          description: "{restaurantId : 변경된 식당의 아이디}"
        403:
          description: '{"message" : "Not Token || invalid token"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "name || floor || lat || lng || category_id || id",
                                    "location": "body"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []
  /restaurant/{id}:
    delete:
      summary: 식당 정보 삭제
      tags:
      - restaurant
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      responses:
        200:
          description: '{"result": true}'
        403:
          description: '{"message" : "Not Token || invalid token"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "id",
                                    "location": "Path Variable"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []

  /review:
    get:
      summary: 리뷰 리스트
      tags:
      - review
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      - in: "body"
        name: "body"
        required: false
        type: object
        properties:
          restaurant_id:
            type: integer
      responses:
        200:
          description: '[{"id": **,"create_datetime": "****-**-** **:**:**","delete_datetime": null,"rating": *,"comment": "*****","user_id": ***,"restaurant_id": ***}]'
        403:
          description: '{"message" : "Not Token || invalid token"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "id",
                                    "location": "body"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []

    post:
      summary: 리뷰 정보 입력
      tags:
      - review
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/review"
      responses:
        200:
          description: "{reviewId : 입력된 리뷰 아이디}"
        403:
          description: '{"message" : "Not Token || invalid token"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "rating || restaurant_id",
                                    "location": "body"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []
        
    put:
      summary: 리뷰 정보 수정
      tags:
      - review
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/review"
      responses:
        200:
          description: "{restaurantId : 변경된 리뷰의 아이디}"
        403:
          description: '{"message" : "Not Token || invalid token"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "rating || id",
                                    "location": "body"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []

  /review/{id}:
    delete:
      summary: 리뷰 정보 삭제
      tags:
      - review
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      responses:
        200:
          description: '{"result": true}'
        403:
          description: '{"message" : "Not Token || invalid token"}'
        422:
          description: 'parameter value 유효성 체크 : {
                            "errors": [
                                {
                                    "value": "***",
                                    "msg": "Invalid value",
                                    "param": "id",
                                    "location": "Path Variable"
                                }
                            ]
                        }'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []

  /tag:
    get:
      summary: 태그 리스트
      tags:
      - tag
      parameters:
      - in: "header"
        name: "x-access-token"
        type: string
      responses:
        200:
          description: "태그 리스트"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/restaurant_category"
        403:
          description: '{"message" : "Not Token || invalid token"}'
        500:
          description: '그 외의 에러 발생시 : {"message" : "Exception"}'
      security:
        - api_key: []
        
definitions:
  user:
    type: object
    properties:
      email:
        type: string
        format: EMAIL

      password:
        type: string
        format: VARCHAR(45)
        
      alias:
        type: string
        format: VARCHAR(45)
        
      phone:
        type: string
        format: VARCHAR(45)
        pattern: '^\d{3}-\d{4}-\d{4}$'

  restaurant:
    type: object
    properties:
      id:
        type: integer
        uniqueItems: true
        
      name:
        type: string
        format: VARCHAR(45)
        
      floor:
        type: integer
        
      introduction:
        type: string
        default: null
        format: TEXT
        
      lat:
        type: number
        format: FLOAT
        
      lng:
        type: number
        format: FLOAT
      
      address:
        type: string
        format: VARCHAR(255)
        default: null
        
      address_road:
        type: string
        format: VARCHAR(255)
        default: null
      
      category_id:
        type: integer
        $ref: '#/definitions/restaurant_category'

  restaurant_category:
    properties:
      id:
        type: integer
        uniqueItems: true
        
      name:
        type: string
        format: VARCHAR(45)
        
  restaurant_image:
    properties:
      id:
        type: integer
        uniqueItems: true
        
      url:
        type: string
        format: VARCHAR(255)
        
      restaurant_id:
        type: integer
        $ref: '#/definitions/restaurant'

  review:
    properties:
      id:
        type: integer
        uniqueItems: true
      
      create_datetime:
        type: string
        format: DATETIME
        
      delete_datetime:
        type: string
        format: DATETIME
        default: null
      
      rating:
        type: number
        format: FLOAT
        
      comment:
        type: string
        format: TEXT
        default: false
        
      user_id:
        type: integer
        $ref: '#/definitions/user'
      
      restaurant_id:
        type: integer
        $ref: '#/definitions/restaurant'
        
  review_image:
    properties:
      id:
        type: integer
        uniqueItems: true
        
      url:
        type: string
        format: VARCHAR(255)
        
      review_id:
        type: integer
        $ref: '#/definitions/review'

  tag:
    properties:
      id:
        type: integer
        uniqueItems: true
        
      name:
        type: string
        format: VARCHAR(45)